feat: remove the dead code of the old business registration service implementation
