+++ server/apps/accounts/services.py
@@ -1,32 +1,59 @@
-from django.contrib.auth import get_user_model
-from django.db.models import Q
-from server.exceptions import ServiceFailed
+from dataclasses import dataclass
+from typing import cast, final
+
+from .dto_types import RawFormData, RawUserCredentials
+from .exceptions import AccountAlreadyExists, MismatchedPasswords, UncorrectPassword
+from .interfaces import PasswordValidator, UserCredentialsRepository
 
 
-__User = get_user_model()
+@final
+@dataclass(frozen=True, slots=True)
+class SignUpService:
+    '''Creates new account.'''
+
+    # Dependencies.
+
+    _storage: UserCredentialsRepository
+    _validate_password: PasswordValidator
 
 
-def create_new_account(raw_form_data: dict[str, str]) -> str:
-    '''
-    Registers a new user in the system.
-    when trying to register an existing account
-    raises a <Service Failed> exception
-    '''
+    def __call__(self, source: RawFormData) -> None:
+        raw_password = self._compare_passwords(source)
+        raw_password = cast(str, raw_password)
+        self._validate_password_strength(raw_password)
+        credentials = self._prepair_user_credentials(source, raw_password)
+        self._check_account_existence(credentials)
+        self._persist_account(credentials)
 
-    filter_condition = Q(
-        username=raw_form_data['username']
-    ) | Q(
-        username__iexact=raw_form_data['username']
-    )
-    if __User.objects.filter(filter_condition).exists():
-        raise ServiceFailed('Attempt to create an existing user')
 
-    new_user = __User(
-        username=raw_form_data['username'],
-        email=raw_form_data['email']
-    )
-    new_user.first_name = raw_form_data['first_name']
-    new_user.last_name = raw_form_data['last_name']
-    new_user.set_password(raw_form_data['password1'])
-    new_user.save()
-    return new_user.first_name
+    def _compare_passwords(self, source_data: RawFormData) -> str | None:
+        password1 = source_data.get('password1')
+        password2 = source_data.get('password2')
+        if password1 != password2:
+            raise MismatchedPasswords(password1, password2)
+        else:
+            return password1
+
+
+    def _validate_password_strength(self, raw_password: str) -> None:
+        try:
+            self._validate_password(raw_password)
+        except ValueError:
+            raise UncorrectPassword(raw_password)
+
+
+    def _prepair_user_credentials(self, source_data: RawFormData, validated_password: str) -> RawUserCredentials:
+        source_data = source_data.copy()
+        del source_data['password1'], source_data['password2']
+        credentials = cast(RawUserCredentials, source_data)
+        credentials['raw_password'] = validated_password
+        return credentials
+
+
+    def _check_account_existence(self, credentials: RawUserCredentials) -> None:
+        if self._storage.contains(credentials):
+            raise AccountAlreadyExists(credentials)
+
+
+    def _persist_account(self, credentials: RawUserCredentials) -> None:
+        self._storage.save(credentials)
